# openapi.yaml
openapi: 3.0.3
info:
  title: Unified OTT API
  version: 1.0.0
servers:
  - url: http://localhost:4000/api
paths:
...
  /content/{id}:
    get:
      summary: Get content details
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
  /content/{id}/recommendations:
    get:
      summary: Related recommendations
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
...